:truck: Rename work estimation component
